pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "crowdtrans"
version = "0.1.0"
description = "Radiology Dictation Transcription Platform (Visage + Karisma)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pymssql>=2.2",
    "deepgram-sdk>=3.7",
    "pydantic-settings>=2.6",
    "click>=8.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
crowdtrans = "crowdtrans.cli:cli"

[tool.setuptools.packages.find]
include = ["crowdtrans*"]
